requirements.txt
allure-behave==2.13.1
allure-pytest==2.13.1
allure-python-commons==2.13.1
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chromedriver-autoinstaller==0.6.2
colorama==0.4.6
decorator==5.1.1
enum34==1.1.10
et-xmlfile==1.1.0
execnet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jsonpath-ng==1.5.3
jsonschema==4.17.3
lxml==4.9.2
namedlist==1.8
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
parse==1.19.0
parse-type==0.6.0
pip==23.2.1
pluggy==1.0.0
ply==3.11
PyAutoIt==0.6.5
pycodestyle==2.10.0
pycparser==2.21
pymssql==2.2.7
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.4.2
pytest-ordering==0.6
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
selenium-page-factory==2.6
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0 
soupsieve==2.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2023.3
urllib3==1.26.14
webdriver-manager==4.0.0
wsproto==1.2.0